+++ CHANGELOG.md
@@ -4,6 +4,13 @@ CHANGELOG ekc-tournament
 ## TODO
 [3] team / tournament: form input/label alignment  
 [3] performance: drop down lists as key / value (integer keys instead of string keys) 
+[2] post-redirect-get and/or get-redirect-get for timer, start round, delete round etc.
+
+# Version 1.2.16
+[x] Bugfix: missing flags on elimination bracket
+[x] Elimination bracket: pre-defined positions in bracket for each rank
+[x] Swiss System: Timer for tie break
+[x] Registration fees: mark paid fees
 
 # Version 1.2.15
 [x] Support for field 'club' in Elementor forms